Move change_value function into a models package
